requirements.txt
absl-py==1.3.0
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
astunparse==1.6.3
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
blinker==1.5
bokeh==3.0.2
bs4==0.0.1
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
compress-pickle==2.1.0
contourpy==1.0.6
cryptography==38.0.3
cycler==0.11.0
decorator==5.1.1
Deprecated==1.2.13
Django==4.1.3
django-appconf==1.0.5
django-crispy-forms==1.14.0
django-imagekit==4.1.0
django-resized==1.0.2
django-starcross-gallery==1.0.14
django-taggit==3.1.0
entrypoints==0.4
exceptiongroup==1.0.4
extra-streamlit-components==0.1.56
Faker==15.3.3
favicon==0.7.0
flatbuffers==22.10.26
fonttools==4.38.0
gast==0.4.0
gcloud==0.18.3
gitdb==4.0.9
GitPython==3.1.29
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.57.0
grpcio==1.50.0
h11==0.14.0
h5py==3.7.0
htbuilder==0.6.1
httplib2==0.21.0
hydralit==1.0.14
hydralit-components==1.0.10
idna==3.4
importlib-metadata==5.0.0
Jinja2==3.1.2
jsonschema==4.17.0
jwcrypto==1.4.2
keras==2.11.0
kiwisolver==1.4.4
libclang==14.0.6
lxml==4.9.1
Markdown==3.4.1
markdownlit==0.0.5
MarkupSafe==2.1.1
matplotlib==3.6.2
more-itertools==9.0.0
numpy==1.23.5
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
pandas==1.5.1
pilkit==2.0
Pillow==9.1.1
protobuf==3.19.6
psycopg2==2.9.5
psycopg2-binary==2.9.1
pyarrow==10.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pydeck==0.8.0
Pygments==2.13.0
pymdown-extensions==9.8
Pympler==1.0.1
pyparsing==3.0.9
Pyrebase4==4.5.0
pyrsistent==0.19.2
PySocks==1.7.1
python-dateutil==2.8.2
python-jwt==4.0.0
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rich==12.6.0
rsa==4.9
selenium==4.7.2
semver==2.13.0
simple-photo-gallery==1.5.4
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.1
st-annotated-text==3.0.0
st-btn-select==0.1.2
streamlit==1.15.1
streamlit-camera-input-live==0.2.0
streamlit-embedcode==0.1.2
streamlit-extras==0.2.3
streamlit-faker==0.0.2
streamlit-keyup==0.2.0
streamlit-login-auth-ui==0.2.0
streamlit-lottie==0.0.3
streamlit-option-menu==0.3.2
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
trio==0.22.0
trio-websocket==0.9.2
trycourier==4.4.0
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
xyzservices==2022.9.0
zipp==3.10.0